delete useless links
