+++ blog/.vuepress/config.js
@@ -1,6 +1,14 @@
 module.exports = {
   title: 'Siren\'s playground',
   description: 'Paint the world colorful.',
+  // title: {
+  //   'en-US': 'Siren\'s playground',
+  //   'zh-CN': 'Siren 的游乐场',
+  // },
+  // description: {
+  //   'en-US': 'Paint the world colorful.',
+  //   'zh-CN': '让世界充满色彩。',
+  // },
   port: 8088,
   markdown: {
     lineNumbers: true,
@@ -15,17 +23,16 @@ module.exports = {
   theme: require.resolve('../../index'), // 使用本地主题
   themeConfig: require('./config/themeConfig'),
   locales: {
-    // 键名是该语言所属的子路径
-    // 作为特例，默认语言可以使用 '/' 作为其路径。
     // '/': {
-    //   lang: 'en-US', // 将会被设置为 <html> 的 lang 属性
-    //   title: 'VuePress',
-    //   description: 'Vue-powered Static Site Generator'
+    //   lang: 'en-US',
+    //   label: 'English',
+    //   title: 'Siren\'s playground',
+    //   description: 'Siren 的游乐场'
     // },
     // '/zh/': {
     //   lang: 'zh-CN',
-    //   title: 'VuePress',
-    //   description: 'Vue 驱动的静态网站生成器'
+    //   label: '简体中文',
+    //   title: '让世界充满色彩',
     // }
   }
 }
