+++ components/Home.vue
@@ -57,25 +57,12 @@
         <div>Of my blogs</div>
         </router-link>
       </div>
-    <div class="about-me">
-      <div class="img-container">
-        <img :src="require('assets/images/selfie.png')">
-      </div>
-      <div class="about-my-info">
-        <div>- A little story about me</div>
-        <!-- <div class="chinese">王欣茹</div> -->
-        <div class="title">I Am <span>Xinru Wang</span></div>
-        <div>{{ selfIntroduction }}</div>
-        <div class="info-box">
-          <div class="info">Birthday: January 12th, 2000</div>
-          <div class="info">Email: siren.xrw@gmail.com</div>
-        </div>
-      </div>
-    </div>
+    <AboutMe />
   </div>
 </template>
 <script>
 import SelectedPostList from '@theme/components/Posts/SelectedPostList.vue';
+import AboutMe from './components/AboutMe.vue'
 import dayjs from 'dayjs'
 import dayjsPluginUTC from 'dayjs/plugin/utc'
 
@@ -85,7 +72,8 @@ const DATE_MAP = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep',
 export default {
   name: 'Home',
   components: {
-    SelectedPostList
+    SelectedPostList,
+    AboutMe
   },
   data() {
     return {
@@ -119,10 +107,7 @@ export default {
           code: 'https://gitee.com/Siren_wang/cyan-mc',
           tags: ['Mobile Application']
         }
-      ],
-      selfIntroduction: `I grew up in the beautiful city of Hangzhou, China. I have always been someone who loves a good challenge and enjoys learning new things. That's why I have decided to pursue my studies abroad.
-I'm also a foodie and a travel enthusiast. I believe that food is not just about nourishment, but also a way to experience cultures and connect with people.
-I really look forward to embark on a new journey and start exploring!`
+      ]
     }
   },
   methods: {
@@ -245,47 +230,6 @@ I really look forward to embark on a new journey and start exploring!`
       
   video
     width: 100%;
-
-.about-me
-  display: flex
-  align-items center
-  background #f8f8f8
-  border-radius 8px
-  margin 60px 0
-  padding 30px
-  .img-container
-    flex: 1
-    text-align center
-    img
-      width 50%
-      transition all .2s ease
-      cursor pointer
-      &:hover
-        filter: drop-shadow(var(--theme-card-boxshadow-hover))
-
-  .about-my-info
-    flex: 1
-    margin-left 20px
-    > div:first-child
-      opacity 0.6
-      // font-size 10px
-    span
-      color: var(--theme-accent-color);
-    .title
-      font-size 2rem
-      font-weight bold
-      margin 14px 0
-      & + div
-        white-space pre-wrap
-    .info-box
-      margin-top 12px
-      display flex
-      flex-wrap wrap
-      .info
-        margin-right 40px
-
-html.dark .about-me
-  background var(--theme-card-background)
 
 h1.block-title
   position: relative;
@@ -413,10 +357,6 @@ h1.block-title
     :not(.article-list) .article-item
       width: 100%;
       margin: 0 !important;
-  .about-me
-    display block
-    .img-container
-      margin-bottom 30px
 
 
 </style>
