+++ components/components/NavLink.vue
@@ -24,7 +24,7 @@
 </template>
 
 <script>
-import { isExternal, isMailto, isTel, ensureExt } from '../util'
+import { isExternal, isMailto, isTel, ensureExt } from '../../util'
 export default {
   name: 'NavLink',
   props: {
